# speed up `roc_curve_survival_impl()`