Before this PR, I saw:

``` r
bench::mark(
  yardstick:::roc_curve_survival_impl(
    truth = rr_churn_data$surv,
    estimate = rr_churn_data$surv_prob,
    censoring_weights = rr_churn_data$ipcw,
    eval_time = rr_churn_data$times
  )
)
#> # A tibble: 1 × 13
#>   expression    min median itr/s…¹ mem_a…² gc/se…³ n_itr  n_gc total…⁴ result
#>   <list>     <bch:> <bch:>   <dbl> <bch:b>   <dbl> <int> <dbl> <bch:t> <list>
#> 1 <chr [1]>   50.6s  50.6s  0.0198   133GB    13.7     1   694   50.6s <tibble>
#> # … with 3 more variables: memory <list>, time <list>, gc <list>, and
#> #   abbreviated variable names ¹​`itr/sec`, ²​mem_alloc, ³​`gc/sec`, ⁴​total_time
#> # ℹ Use `colnames()` to see all variable names
```

With this PR:

``` r
#> # A tibble: 1 × 13
#>   expression    min median itr/s…¹ mem_a…² gc/se…³ n_itr  n_gc total…⁴ result
#>   <bch:expr> <bch:> <bch:>   <dbl> <bch:b>   <dbl> <int> <dbl> <bch:t> <list>
#> 1 total       7.59s  7.59s   0.132  12.7GB    40.1     1   304   7.59s <tibble>
#> # … with 3 more variables: memory <list>, time <list>, gc <list>, and
#> #   abbreviated variable names ¹​`itr/sec`, ²​mem_alloc, ³​`gc/sec`, ⁴​total_time
#> # ℹ Use `colnames()` to see all variable names
```

@EmilHvitfeldt mentioned there could be a `sort()`ing trick we could use for the `<=` comparison that might drive this down further. :)