@simonpcouch 2023-03-17
Related to #379! Maybe closes.