@EmilHvitfeldt 2023-03-17
I hit it with the good old 

```r
new_res <- yardstick:::roc_curve_survival_impl(
  truth = rr_churn_data$surv,
  estimate = rr_churn_data$surv_prob,
  censoring_weights = rr_churn_data$ipcw,
  eval_time = rr_churn_data$times
)

# readr::write_rds(true_res, "~/Desktop/old_res.rds")
old_res <- readr::read_rds("~/Desktop/old_res.rds")

identical(new_res, old_res)
```

and we are still good!