@EmilHvitfeldt 2023-03-17
We are now at 

```r
#> # A tibble: 1 × 13
#>   expression     min median itr/s…¹ mem_a…² gc/se…³ n_itr  n_gc total…⁴ result  
#>   <bch:expr> <bch:t> <bch:>   <dbl> <bch:b>   <dbl> <int> <dbl> <bch:t> <list>  
#> 1 total         5.6s   5.6s   0.179  9.51GB    38.0     1   213    5.6s <tibble>
#> # … with 3 more variables: memory <list>, time <list>, gc <list>, and
#> #   abbreviated variable names ¹​`itr/sec`, ²​mem_alloc, ³​`gc/sec`, ⁴​total_time
#> # ℹ Use `colnames()` to see all variable names
```