@EmilHvitfeldt 2023-03-18
We are now finally down to an acceptable amount of time to run.

```r
#> # A tibble: 1 × 13
#>   expression     min median itr/s…¹ mem_a…² gc/se…³ n_itr  n_gc total…⁴ result  
#>   <bch:expr> <bch:t> <bch:>   <dbl> <bch:b>   <dbl> <int> <dbl> <bch:t> <list>  
#> 1 total        359ms  359ms    2.79  11.3MB    13.9     2    10   717ms <tibble>
#> # … with 3 more variables: memory <list>, time <list>, gc <list>, and
#> #   abbreviated variable names ¹​`itr/sec`, ²​mem_alloc, ³​`gc/sec`, ⁴​total_time
#> # ℹ Use `colnames()` to see all variable names
```

I'll need to do some more testing to make sure everything is correct. Since we are sorting things, some sums are now slightly different (to be expected) but I want to make sure it is correct

![Screen Shot 2023-03-17 at 3 38 36 PM](https://user-images.githubusercontent.com/14034784/226074325-9aa61d05-0d3c-4a44-9892-d83ca1a42872.png)